fix(gophermart): rename customerrors for tests
